+++ resources/assets/js/components/FrontContact.vue
@@ -1,0 +1,8 @@
+<template>
+<div id="FrontContact" class="container">
+    <div class="p-md-3 col-9 col-sm-9 col-md-9 mx-auto">
+        <h2>コンタクト</h2>
+        <hr>
+    </div>
+</div>
+</template>
