+++ resources/assets/js/components/FrontHeader.vue
@@ -4,30 +4,15 @@
         <h1 id="SiteTitle"><router-link :to="'/'">{{ title }}</router-link></h1>
         <p>{{ description }}</p>
         <ul id="Menu">
-            <li class="menu-item"><router-link :to="'/'">Profile</router-link></li>
-            <li class="menu-item"><router-link :to="'/'">Contact</router-link></li>
+            <li class="menu-item"><router-link :to="'/profile'">Profile</router-link></li>
+            <li class="menu-item"><router-link :to="'/contact'">Contact</router-link></li>
         </ul>
         <hr class="col-5 col-sm-5 col-md-2 text-center">
-        <div id="Search" class="mb-4">
-            <div class="col-4 col-sm-4 col-md-4 mx-auto">
-                <input type="text" class="form-control text-center p-0" placeholder="しぼりこみけんさく" v-on:keyup.enter="pushFilteredIndex">
-            </div>
-        </div>
     </header>
 </div>
 </template>
 <script>
 export default {
-    props: ['title', 'description'],
-    methods: {
-        pushFilteredIndex(event) {
-            const keyword = encodeURIComponent(event.target.value)
-            if (keyword) {
-                this.$router.push({ path: '/?keyword=' + keyword })
-            } else {
-                this.$router.push({ path: '/' })
-            }
-        }
-    }
+    props: ['title', 'description']
 }
 </script>
