+++ resources/assets/js/components/FrontProfile.vue
@@ -1,0 +1,8 @@
+<template>
+<div id="FrontProfile" class="container">
+    <div class="p-md-3 col-9 col-sm-9 col-md-9 mx-auto">
+        <h2>プロフィール</h2>
+        <hr>
+    </div>
+</div>
+</template>
